Bump timeout to 20ms
Apparently the timer resolution isn't fine enough; I've observed this value as
high as 22ms, but I just ran the tests ~100 times with the diff set to 20ms.

If it's a problem I can bump higher or we can try to figure out some way to
force the ordering here with fake timers.